Struct darwin_webkit::helpers::dwk_app::DarwinWKApp

pub struct DarwinWKApp { /* fields omitted */ }

Wraps an NSApplication instance with a main window that contains WebView.

See DarwinWKWebView as well.

Example

use darwin_webkit::helpers::dwk_app::DarwinWKApp;
use std::rc::Rc;

unsafe {
    let app = DarwinWKApp::new("Host an app");
    let webview = Rc::new(app.create_webview());

    // add handlers, load HTML, etc...

    app.set_webview(&webview);
    app.run();
}

Implementations

impl DarwinWKApp

pub unsafe fn new(windowTitle: &str) -> DarwinWKApp

Create an app with the given window title

Safety

All the FFI functions are unsafe.

pub fn get_app_native_handle(&self) -> id

Get the NSApplication handle

pub fn get_window_native_handle(&self) -> id

Get the NSWindow handle

pub unsafe fn run(&self)

Start the NSApplication and activate it ignoring other apps so it comes to front.

Safety

All the FFI functions are unsafe.

pub unsafe fn stop(&self)

Stop the NSApplication run loop.

Safety

All the FFI functions are unsafe.

pub unsafe fn create_webview(&self) -> DarwinWKWebView

Create a webview that has this app window’s frame.

Safety

All the FFI functions are unsafe.

pub unsafe fn set_webview<'a>(&'a self, webview: &'a DarwinWKWebView)

Set the content view of the main window to a certain webview

Safety

All the FFI functions are unsafe.